feat(bulk): underlying different implementations in api and core of search
Refs: #1793
