+++ packages/api/src/external/fhir/document/index.ts
@@ -58,22 +58,6 @@ const authorTypesMap: Record<AuthorTypes["resourceType"], AuthorTypes["resourceT
 };
 const authorTypes = Object.values(authorTypesMap);
 
-export function isDocStatusReady(doc: DocumentReference): boolean {
-  return !isDocStatusPreliminary(doc) && !isDocStatusEnteredInError(doc);
-}
-export function isDocStatusFinal(doc: DocumentReference): boolean {
-  return doc.docStatus === "final";
-}
-export function isDocStatusAmended(doc: DocumentReference): boolean {
-  return doc.docStatus === "amended";
-}
-export function isDocStatusPreliminary(doc: DocumentReference): boolean {
-  return doc.docStatus === "preliminary";
-}
-export function isDocStatusEnteredInError(doc: DocumentReference): boolean {
-  return doc.docStatus === "entered-in-error";
-}
-
 // HIEs probably don't have records before the year 1800 :)
 const earliestPossibleYear = 1800;
