+++ packages/core/src/external/opensearch/search-documents.ts
@@ -1,10 +1,12 @@
 import { DocumentReference } from "@medplum/fhirtypes";
+import { isCommonwellExtension } from "../commonwell/extension";
+import { isCarequalityExtension } from "../carequality/extension";
+import { isMetriportExtension } from "../fhir/shared/extensions/metriport";
 import { uniqBy } from "lodash";
+import { isDocStatusReady } from ".";
 import { Config } from "../../util/config";
 import { capture } from "../../util/notifications";
-import { isCommonwellExtension } from "../commonwell/extension";
 import { makeSearchServiceQuery } from "../opensearch/file-search-connector-factory";
-import { isMetriportExtension } from "../fhir/shared/extensions/metriport";
 import { getDocuments } from "../fhir/document/get-documents";
 
 export async function searchDocuments({
@@ -33,7 +35,8 @@ export async function searchDocuments({
   }
 
   const unique = uniqBy(success, "id");
-  return unique;
+  const ready = unique.filter(isDocStatusReady);
+  return ready;
 }
 
 async function searchOnDocumentReferences(
@@ -54,7 +57,8 @@ function checkExtensions(doc: DocumentReference) {
   if (!extensions) return false;
   const metriport = extensions.find(isMetriportExtension);
   const cw = extensions.find(isCommonwellExtension);
-  if (!metriport && !cw) return false;
+  const cq = extensions.find(isCarequalityExtension);
+  if (!metriport && !cw && !cq) return false;
   return true;
 }
